first batch of sql tests
