+++ python/samples/concepts/memory/complex_memory.py
@@ -29,7 +29,7 @@ from semantic_kernel.connectors.memory.pinecone import PineconeCollection
 from semantic_kernel.connectors.memory.postgres import PostgresCollection
 from semantic_kernel.connectors.memory.qdrant import QdrantCollection
 from semantic_kernel.connectors.memory.redis import RedisHashsetCollection, RedisJsonCollection
-from semantic_kernel.connectors.memory.sql_server import AzureSqlCollection
+from semantic_kernel.connectors.memory.sql_server import SqlServerCollection
 from semantic_kernel.connectors.memory.weaviate import WeaviateCollection
 from semantic_kernel.data import (
     VectorizableTextSearchMixin,
@@ -148,6 +148,9 @@ DataModel = get_data_model("array", index_kind, distance_function)
 #   The chroma collection is currently only available for in-memory versions
 #   Client-Server mode and Chroma Cloud are not yet supported.
 #   More info on Chroma here: https://docs.trychroma.com/docs/overview/introduction
+# - faiss: Faiss - in-memory with optimized indexes.
+# - pinecone: Pinecone
+# - sql_server: SQL Server, can connect to any SQL Server compatible database, like Azure SQL.
 # This is represented as a mapping from the collection name to a
 # function which returns the collection.
 # Using a function allows for lazy initialization of the collection,
@@ -205,7 +208,7 @@ collections: dict[str, Callable[[], VectorStoreRecordCollection]] = {
         collection_name=collection_name,
         data_model_type=DataModel,
     ),
-    "azure_sql": lambda: AzureSqlCollection[str, DataModel](
+    "sql_server": lambda: SqlServerCollection[str, DataModel](
         data_model_type=DataModel,
         collection_name=collection_name,
     ),
@@ -344,5 +347,5 @@ if __name__ == "__main__":
     # Option of whether to use OpenAI or Azure OpenAI.
     parser.add_argument("--use-azure-openai", action="store_true", help="Use Azure OpenAI instead of OpenAI.")
     args = parser.parse_args()
-    args.collection = "azure_sql"
+    args.collection = "sql_server"
     asyncio.run(main(collection=args.collection, use_azure_openai=args.use_azure_openai))
