--- python/semantic_kernel/connectors/memory/sql_server/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from semantic_kernel.connectors.memory.sql_server._azure_sql_server import AzureSqlCollection
-
-__all__ = ["AzureSqlCollection"]
